backend/requirements.txt
flask==3.0.3
gunicorn==23.0.0
mysql-connector-python==9.0.0
requests==2.31.0
